feat: :zap: Layout of the left section.
I designed the structure of the right section, which I consider easier to tackle. However, when trying to position the different boxes with flexbox, it seemed unnecessary to me. On the other hand, as usual, I created my test files to store the failed attempts and use them as a reference.